// src/redux/ducks/companies/types.ts
export type MetaInformationProps = {
  currentPage: string;
  itemCount: number;
  itemsPerPage: string;
  totalItems: number;
  totalPages: number;
};

export type ItemsProps = {
  id: string;
  zoomInfoId?: any;
  name: string;
  logo?: any;
  ticker?: any;
  parentCompany?: any;
  phone?: string;
  fax?: string;
  website?: string;
  city?: string;
  street?: string;
  state?: string;
  zipCode?: string;
  country?: string;
  continent?: any;
  productsBrandDescription?: any;
  descriptionList?: string;
  revenueRange?: string;
  employeeRange?: string;
  twitterHandle?: any;
  socialMediaUrls?: any;
  competitors?: any;
  subUnitIndustries?: any;
  primaryIndustry?: string[];
  industries?: any;
  revenue?: string;
  employeeCount?: number;
  annualContributions?: any;
  cashContributions?: any;
  inKindContributions?: any;
  employeeContributions?: any;
  parentId?: string;
  parentName?: any;
  type?: any;
  sdgGoals?: any;
  genders?: any;
  income?: any;
  age?: any;
  ethnicity?: any;
  nonprofit?: any;
  purchase?: any;
  affiliation?: any;
  brands?: any;
  interests?: any;
  typesOfInvestment?: any;
  errorLoadZoomInfo?: any;
  charitablePartners?: any;
  statusZoomInfo?: string;
  loadZoomInfoDate?: any;
  errorLoadZoomInfoDate?: any;
  partnershipLink?: any;
  employeeEngagementOpportunities?: true;
  similarCompanies?: string[];
  favoriteCompanies?: any;
  score?: number;
  like?: boolean;
  crsFocus?: any;
};